+++ src/main/resources/templates/FilmRouletten/index.html
@@ -9,7 +9,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Xanh+Mono:ital,wght@0,400;1,400&display=swap" rel="stylesheet" />
 
     <link href="../../public/css/styles.css" th:href="@{/css/styles.css}" rel="stylesheet" />
-    <link href="../../public/css/FilmRouletten.css" th:href="@{/css/FilmRouletten.css}" rel="stylesheet" />
+    <link href="../../public/css/filmRouletten/FilmRouletten.css" th:href="@{/css/filmRouletten/FilmRouletten.css}" rel="stylesheet" />
 </head>
 
 <body>
